requirements.txt
yfinance
feedparser
pyyaml
python-dotenv
pandas
requests
beautifulsoup4
schedule
langchain-google-genai
